Added a package management for importing/exporting excel/csv files for students records

+++ resources/views/admin/manage-students.blade.php
@@ -28,6 +28,8 @@
                             Name</th>
                         <th class="px-6 py-3 text-xs font-medium leading-4 tracking-wider text-left text-gray-500 uppercase border-b border-gray-200 bg-gray-50">
                             Reg. Number</th>
+                            <th class="px-6 py-3 text-xs font-medium leading-4 tracking-wider text-left text-gray-500 uppercase border-b border-gray-200 bg-gray-50">
+                            Date of Birth</th>
                         <th class="px-6 py-3 text-xs font-medium leading-4 tracking-wider text-left text-gray-500 uppercase border-b border-gray-200 bg-gray-50">
                             Status</th>
                         <th class="px-6 py-3 text-xs font-medium leading-4 tracking-wider text-left text-gray-500 uppercase border-b border-gray-200 bg-gray-50">
@@ -39,7 +41,7 @@
 
                 <tbody class="bg-white">
                     <tr>
-                        @php 
+                        @php
 
                         $i = 1;
 
@@ -67,18 +69,25 @@
 
                         </td>
 
+                        <td class="px-6 py-4 whitespace-no-wrap border-b border-gray-200">
+                            <div class="text-sm leading-5 text-gray-500">{{ $student->dob }}</div>
+
+                        </td>
+
+
 
                         <td class="px-6 py-4 whitespace-no-wrap border-b border-gray-200">
-                            <span class="inline-flex px-2 text-xs font-semibold leading-5 text-green-800 bg-green-100 rounded-full">Active</span>
+                            <span class="inline-flex px-2 text-xs font-semibold leading-5 text-green-800 bg-green-300 rounded-full">Active</span>
                         </td>
 
                         <td class="px-6 py-4 text-sm leading-5 text-gray-500 whitespace-no-wrap border-b border-gray-200">
-                         <a href="#" class="px-4 py-1 text-sm text-white bg-green-400 rounded-full">Edit</a>
+                            <a href="#" class="px-4 py-1 text-sm text-white bg-green-400 rounded-full">Edit</a>
                         </td>
 
                         <td class="px-6 py-4 text-sm leading-5 text-gray-500 whitespace-no-wrap border-b border-gray-200">
-                         <a href="/deleteStudent/{{ $student->id }}" class="px-4 py-1 text-sm text-black bg-red-500 rounded-full">Delete</a>
+                            <a href="/deleteStudent/{{ $student->id }}" class="px-4 py-1 text-sm text-black bg-red-500 rounded-full" onclick="confirm('Are you sure you want to delete?')">Delete</a>
                         </td>
+
                     </tr>
 
                     <tr>
@@ -90,7 +99,19 @@
 
 
             </table>
-        </div>  
+        </div>
     </div>
 </div>
 {{ $students->links() }}
+
+<script>
+function deleteConfirmation() {
+  var txt;
+  if (confirm("Press a button!")) {
+    txt = "You pressed OK!";
+  } else {
+    txt = "You pressed Cancel!";
+  }
+  document.getElementById("demo").innerHTML = txt;
+}
+</script>
